+++ frontend/src/strava.rs
@@ -21,7 +21,7 @@ enum CompleteDownload {
 #[derive(PartialEq)]
 pub enum LoadState {
     /// Continue loading. Includes the next after_id.
-    Continue(Option<i32>),
+    Continue(Option<i64>),
     /// Finished loading.
     Finished,
 }
@@ -84,7 +84,7 @@ async fn fetch_json<T: DeserializeOwned>(
 /// `after_id` pages through results: only runs with a `strava_activity_id`
 /// at or beyond it are returned by the backend.
 async fn fetch_runs(
-    after_id: Option<i32>,
+    after_id: Option<i64>,
 ) -> Result<(Vec<comms::runs::RunResponse>, CompleteDownload), LoadError> {
     let mut url = format!("{BACKEND_BASE_URL}/api/runs");
     if let Some(after_id) = after_id {
@@ -107,7 +107,7 @@ fn decode_line(encoded: &str) -> Vec<Vec<f64>> {
 /// Fetch recent runs and return them as a GeoJSON `FeatureCollection` of `LineString`s.
 ///
 /// Pass `after_id` to fetch the following page or pass `None` for the initial load.
-pub async fn load_run_lines(after_id: Option<i32>) -> Result<LoadedRuns, LoadError> {
+pub async fn load_run_lines(after_id: Option<i64>) -> Result<LoadedRuns, LoadError> {
     let (runs, complete_download) = fetch_runs(after_id).await?;
 
     let features = runs
